Skeleton for all test cases
