--- test/unit/engine/pandas/test_is_in_billions.py
@@ -1,18 +1,0 @@
-from .base import PandasTestCase
-
-
-class TestCase(PandasTestCase):
-    def test_pass(self):
-        assert True
-
-    def test_fail(self):
-        assert True
-
-    def test_exception(self):
-        assert True
-
-    def test_parameter(self):
-        assert True
-
-    def test_coverage(self):
-        assert True
