--- test/unit/engine/pyspark/test_is_less_or_equal_than.py
@@ -1,18 +1,0 @@
-from .base import PysparkTestCase
-
-
-class TestCase(PysparkTestCase):
-    def test_pass(self):
-        assert True
-
-    def test_fail(self):
-        assert True
-
-    def test_exception(self):
-        assert True
-
-    def test_parameter(self):
-        assert True
-
-    def test_coverage(self):
-        assert True
